refactor(tests): move performance tests to testInteg/perf/ #5735
## Problem
Performance test involve running the same code 10 times, and often its
demanding code. This should live with integ rather than unit tests to
avoid slow down.

## Solution
- Move
https://github.com/aws/aws-toolkit-vscode/blob/master/packages/core/src/test/amazonqFeatureDev/prepareRepoData.test.ts
into integ folder.
- Move
https://github.com/aws/aws-toolkit-vscode/blob/master/packages/core/src/test/codewhisperer/commands/startSecurityScan.test.ts
into integ folder.
